requirements.txt
streamlit==1.31.1
PyMuPDF==1.23.21
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.38.2
langchain==0.1.7